requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
numpy
opt-einsum
email-validator
opencv_contrib_python
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paho_mqtt==1.6.1
pandas==2.2.3
psycopg2==2.9.9
psycopg2_binary==2.9.9
PyYAML
supervision==0.23.0
ultralytics==8.2.84
WTForms==3.1.2
